requirements.txt
fastapi
uvicorn
chromadb
openai
langgraph
langchain
langchain-openai
langchain-community
pydantic
python-multipart
python-dotenv
numpy
tiktoken
sentence-transformers
typing-extensions
jq
faiss-cpu
langchain_google_genai